Move TestAgora to Utils for reusability
This will allow us to reuse this instance once we depend
on Agora being up when starting.
